Most stages are lazily loaded now
