+++ examples/gnn_fraud_detection_pipeline/stages/__init__.py
@@ -1,3 +1,3 @@
-from . import classification_stage
+# from . import classification_stage
 from . import graph_construction_stage
 from . import graph_sage_stage
