+++ morpheus/stages/general/buffer_stage.py
@@ -16,6 +16,7 @@ import logging
 import typing
 
 import srf
+from morpheus.cli.register_stage import register_stage
 
 from morpheus.config import Config
 from morpheus.pipeline.single_port_stage import SinglePortStage
@@ -25,8 +26,11 @@ from morpheus.utils.logging import deprecated_stage_warning
 logger = logging.getLogger(__name__)
 
 
+@register_stage("buffer")
 class BufferStage(SinglePortStage):
     """
+    Buffer results.
+
     The input messages are buffered by this stage class for faster access to downstream stages. Allows
     upstream stages to run faster than downstream stages.
