+++ morpheus/stages/output/write_to_kafka_stage.py
@@ -19,6 +19,7 @@ import typing
 import confluent_kafka as ck
 import srf
 from srf.core import operators as ops
+from morpheus.cli.register_stage import register_stage
 
 from morpheus.config import Config
 from morpheus.io import serializers
@@ -29,9 +30,10 @@ from morpheus.pipeline.stream_pair import StreamPair
 logger = logging.getLogger(__name__)
 
 
+@register_stage("to-kafka")
 class WriteToKafkaStage(SinglePortStage):
     """
-    Write messages to a Kafka cluster.
+    Write all messages to a Kafka cluster.
 
     Parameters
     ----------
