+++ morpheus/stages/postprocess/ml_flow_drift_stage.py
@@ -19,8 +19,9 @@ import typing
 import cupy as cp
 import mlflow
 import srf
+from morpheus.cli.register_stage import register_stage
 
-from morpheus.config import Config
+from morpheus.config import Config, PipelineModes
 from morpheus.messages import MultiResponseMessage
 from morpheus.messages import MultiResponseProbsMessage
 from morpheus.pipeline.single_port_stage import SinglePortStage
@@ -29,8 +30,11 @@ from morpheus.pipeline.stream_pair import StreamPair
 logger = logging.getLogger(__name__)
 
 
+@register_stage("mlflow-drift", modes=[PipelineModes.FIL, PipelineModes.NLP, PipelineModes.OTHER])
 class MLFlowDriftStage(SinglePortStage):
     """
+    Report model drift statistics to ML Flow.
+
     Caculates model drift over time and reports the information to MLflow.
 
     Parameters
