+++ morpheus/stages/preprocess/drop_null_stage.py
@@ -17,15 +17,18 @@ import typing
 import srf
 from srf.core import operators as ops
 
+from morpheus.cli.register_stage import register_stage
 from morpheus.config import Config
+from morpheus.config import PipelineModes
 from morpheus.messages import MessageMeta
 from morpheus.pipeline.single_port_stage import SinglePortStage
 from morpheus.pipeline.stream_pair import StreamPair
 
 
+@register_stage("dropna", modes=[PipelineModes.FIL, PipelineModes.NLP, PipelineModes.OTHER])
 class DropNullStage(SinglePortStage):
     """
-    Drop null/empty data input entries.
+    Drop null data entries from a DataFrame.
 
     Parameters
     ----------
