+++ morpheus/utils/type_utils.py
@@ -126,3 +126,10 @@ def mirror_args(wrapped: _DecoratorType,
                 assigned=('__doc__', '__annotations__'),
                 updated=functools.WRAPPER_UPDATES) -> typing.Callable[[_DecoratorType], _DecoratorType]:
     return functools.wraps(wrapped=wrapped, assigned=assigned, updated=updated)
+
+
+def get_full_qualname(klass: typing.Type) -> str:
+    module = klass.__module__
+    if module == '__builtin__':
+        return klass.__qualname__
+    return module + '.' + klass.__qualname__
